# api-docs/swagger.yaml
basePath: /
definitions:
...
  httpserver.errorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  httpserver.errorEnvelope:
    properties:
      error:
        $ref: '#/definitions/httpserver.errorBody'
      success:
        example: false
        type: boolean
    type: object
info:
  contact: {}
  description: 私有部署企业协作产品的服务端 API。当前阶段包含管理员登录、管理员创建普通用户、普通用户登录。
  title: AI 原生企业协作服务 API
  version: 0.1.0
paths:
...
  /api/app/ws:
    get:
      description: 应用使用 App ID 和连接密钥连接，连接存在且心跳正常时视为在线。
      parameters:
      - description: 应用 ID
        in: header
        name: X-MagicChat-App-ID
        required: true
        type: string
      - description: Bearer 连接密钥
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpserver.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpserver.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpserver.errorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpserver.errorEnvelope'
      summary: 应用 WebSocket 连接
      tags:
      - 应用连接
...
securityDefinitions:
  UserSession:
    description: 使用 user_session=<token> 格式的会话 Cookie。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"